[#683] move Resource utils into resources module
