+++ android/src/main/kotlin/com/oztechan/ccc/android/ui/currencies/CurrenciesAdapter.kt
@@ -5,9 +5,9 @@ import android.view.ViewGroup
 import android.view.animation.AnimationUtils
 import androidx.recyclerview.widget.DiffUtil
 import com.github.submob.basemob.adapter.BaseVBRecyclerViewAdapter
-import com.oztechan.ccc.android.util.setBackgroundByName
 import com.oztechan.ccc.client.model.Currency
 import com.oztechan.ccc.client.viewmodel.currencies.CurrenciesEvent
+import com.oztechan.ccc.resources.setBackgroundByName
 import mustafaozhan.github.com.mycurrencies.R
 import mustafaozhan.github.com.mycurrencies.databinding.ItemCurrenciesBinding
