+++ android/src/main/kotlin/com/oztechan/ccc/android/ui/selectcurrency/SelectCurrencyAdapter.kt
@@ -4,9 +4,9 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import com.github.submob.basemob.adapter.BaseVBRecyclerViewAdapter
-import com.oztechan.ccc.android.util.setBackgroundByName
 import com.oztechan.ccc.client.model.Currency
 import com.oztechan.ccc.client.viewmodel.selectcurrency.SelectCurrencyEvent
+import com.oztechan.ccc.resources.setBackgroundByName
 import mustafaozhan.github.com.mycurrencies.databinding.ItemSelectCurrencyBinding
 
 class SelectCurrencyAdapter(
