+++ android/src/main/kotlin/com/oztechan/ccc/android/util/ViewExtensions.kt
@@ -14,35 +14,16 @@ import android.view.animation.Animation
 import android.view.animation.Transformation
 import android.view.inputmethod.InputMethodManager
 import android.widget.FrameLayout
-import android.widget.ImageView
 import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import co.touchlab.kermit.Logger
-import com.github.submob.logmob.e
 import com.github.submob.scopemob.castTo
 import com.oztechan.ccc.ad.AdManager
 import com.oztechan.ccc.client.model.RateState
 import mustafaozhan.github.com.mycurrencies.R
-import java.io.FileNotFoundException
 
 private const val ANIMATION_DURATION = 500L
-
-fun ImageView.setBackgroundByName(
-    name: String
-) = setImageResource(context.getImageResourceByName(name))
-
-fun Context.getImageResourceByName(name: String): Int = try {
-    resources.getIdentifier(
-        name.lowercase().replace("try", "tryy"),
-        "drawable",
-        packageName
-    )
-} catch (e: FileNotFoundException) {
-    Logger.e(e)
-    R.drawable.unknown
-}
 
 fun View.hideKeyboard() = context?.getSystemService(Context.INPUT_METHOD_SERVICE)
     ?.castTo<InputMethodManager>()
