+++ resources/src/androidMain/kotlin/com/oztechan/ccc/resources/AndroidResources.kt
@@ -1,0 +1,12 @@
+package com.oztechan.ccc.resources
+
+import android.widget.ImageView
+
+fun getDrawableIdByFileName(
+    name: String
+): Int = getImageByFileName(name.toImageFileName()).drawableResId
+
+fun ImageView.setBackgroundByName(
+    name: String
+) = setImageResource(getDrawableIdByFileName(name))
+
